+++ probleme493.cpp
@@ -8,7 +8,7 @@ ENREGISTRER_PROBLEME(493, "Under The Rainbow") {
     // What is the expected number of distinct colors in 20 randomly picked balls?
     //
     // Give your answer with nine digits after the decimal point (a.bcdefghij).
-    long double resultat = 7 * (1.0l - mp_nombre::coefficient_binomial(60u, 20u).get_double() 
+    double resultat = 7 * (1.0 - mp_nombre::coefficient_binomial(60u, 20u).get_double()
         / mp_nombre::coefficient_binomial(70u, 20u).get_double());
     return std::to_string(resultat, 9);
 }
